Region transition
The dungeons can now be entered and the transition between regions is now in place. Potentially we can create infinite regions that can be transitioned to and from. Each region is specified with details set in an Object (called world).
The dungeon tiles has been created with floor tiles and a pillar obstacle as well as a couple of barrels as another obstacle.
We can also define specific types of enemies for each world (to be generated randomly an scattered throughout the world)
Tooltips has been updated to look better and also to show/handle dynamic content.
Added a visual border for when the player reaches the outer perimeters in the current region.