Inventory, daytime cycle, battlesystem
Inventory has been updated quite a bit. Weapons and armor are placed in the inventory when collected and you can choose which item to equip (if equipable). Also weapons and armor are now swapped back into inventory when replaced.
Daytime cycle added. Daytime- and nighttime- length is adjustable and the player can light a torch when visibility is too low. Used requestAnimationFrame() to control daytime/nightime transitions smoothly (this was nescessary for the torch on/off feature). I also had to redo the tooltips used across different elements, because of the nighttime overlay, which created conflicts with z-index between tiles, tooltips and the overlay.
Added actionpointcost to weapons and armor to be calculated when moving and attacking in battle (part of the battlesystem).
A small popup menu has been added (toggled with "e" -key). With the menu you can choose to open the inventory, the map and the torch (more is likely to be added to this menu).