Home added and fixed gateways
Added a home (house) for the player and fixed some issues with transitioning between worlds/regions (between open world/home/dungeons/etc.).
Added a sun image that rises and lowers in sync with the daytime cycle.
Adjusted the amount of generated loot to a percentage of the size of the region (instead of a fixed amount)