Spells added + numerous small fixes
New spells added:
- Fireball, ranged with burning effect (DOT)
- Ice Nova, hit all surrounding enemies
- Magic Shield, protect yourself from incoming damage

Spell crafting feature:
Collect gems to improve spells:
- Red gems: improve damage/protection (armor)
- Blue gems: lower actionpoint cost
- Green gems: lower cooldown
(This still needs a lot of adjustments)

The original intention for movement was to use keyboard only (arrow keys or WASD), but made the choice to change this, so now the main movement and navigation is made with the mouse (you can still move with arrow keys and WASD).

Added a flee function, so you can escape a fight if needed (currently there is a cost of 5 AP to do this)

Added new chest types with different amount of loot in it (lucky find?)
Also added a "Take all" button to fully loot a chest  in one go. When the chest is empty, the dialog closes and the chest self destructs.