Quests added
This was propably the most difficult for me to create, but I managed to create a quest system, that I can configure as I please.
Unlimited number of quests possible.
Unlimited number of quest chains (as many sub objectives needed for the whole quest to be completed)
Reward can be anything thanks to callbacks.
Quest objective can be anything (I think). Currently I have only tested with a "delivery" -quest and "kill a certain number of specific enemies" -quest.
UI for dialog (unlimited number of dialogs per quest, automatically displays a next -button until the last dialog is displayed)
UI for quest destination. An exlamation mark ( ! ) is displayed at the destination for the quest (both on the minimap and at the actual destination in the zone) and they update when the quest updates.