Spells and quests
Spells updated with a fire button for the "no-target" spells.
Added an end turn button.
Adjusted action points a bit.
Tweaked the effects a bit.

Quest overview window updated. Now has a sort button to sort between ALL / NOT COMPLETED / COMPLETED.
Clicking a row of a quest, shows the previously displayed dialogs for that quest.

Changed the position of the messages in the messaging system.